cloud_functions/cf-pdf2bq/requirements.txt
google-cloud-secret-manager==2.22.0
google-cloud-storage==2.18.0
google-cloud-bigquery==3.26.0
google-cloud-logging==3.8.0
langchain==0.3.13
langchain-google-genai==2.0.7
backoff==2.2.1
langchain-google-vertexai==2.0.9